update(3)

+++ app/Http/Controllers/DashboardReservasiController.php
@@ -4,9 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\Reservasi;
 use Illuminate\Http\Request;
+use App\Exports\PreservasiExport;
+use Maatwebsite\Excel\Facades\Excel;
 
 class DashboardReservasiController extends Controller
 {
+    public function export(){
+        return Excel::download(new PreservasiExport, 'data_preservasi.xlsx');
+    }
+    
     /**
      * Display a listing of the resource.
      *
@@ -30,6 +36,8 @@ class DashboardReservasiController extends Controller
      */
     public function create()
     {
+        $this->authorize('admin');
+        
         return view('dashboard.preservasi.create');
     }
 
@@ -41,6 +49,8 @@ class DashboardReservasiController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('admin');
+
         $validatedData = $request->validate([
             'tanggal' => 'required',
             'nama_kegiatan' => 'required|max:255',
@@ -92,6 +102,8 @@ class DashboardReservasiController extends Controller
      */
     public function edit(Reservasi $preservasi)
     {
+        $this->authorize('admin');
+
         return view('dashboard.preservasi.edit',[
             // dd($preservasi),
             'preservasi' => $preservasi
@@ -107,6 +119,8 @@ class DashboardReservasiController extends Controller
      */
     public function update(Request $request, Reservasi $preservasi)
     {
+        $this->authorize('admin');
+
         $validatedData = $request->validate([
             'tanggal' => 'required',
             'nama_kegiatan' => 'required|max:255',
@@ -139,7 +153,8 @@ class DashboardReservasiController extends Controller
      */
     public function destroy(Reservasi $preservasi)
     {
-        
+        $this->authorize('admin');
+
         Reservasi::destroy($preservasi->id);
         return redirect('/dashboard/preservasi')->with('success', 'Data sudah dihapus');
